+++ packages/hooks/index.js
@@ -3,3 +3,4 @@ export { default as useMount } from './src/useMount';
 export { default as useTimeout } from './src/useTimeout';
 export { default as useToggle } from './src/useToggle';
 export { default as useCurrentRegion } from './src/useCurrentRegion';
+export { default as useCurrentUser } from './src/useCurrentUser';