+++ packages/upload/tests/UploadProgressBar.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '@testing-library/react';
+import { render, cleanup } from '@testing-library/react';
 import MockUpload from './mockUpload';
 import { UploadProgressBar } from '..';
 
@@ -8,6 +8,8 @@ const instance = new MockUpload();
 describe('UploadProgressBar', () => {
   afterEach(() => {
     instance.reset();
+    cleanup();
+    jest.clearAllMocks();
   });
   test('should render', () => {
     render(<UploadProgressBar upload={instance} />);
@@ -47,33 +49,34 @@ describe('UploadProgressBar', () => {
   // striped
   // animated
 
-  test('should render striped', () =>{
-    const { getByTestId } = render(<UploadProgressBar upload={instance} striped />);
-    
-    const progressBar = getByTestId('progress-inner')
+  test('should render striped', () => {
+    const { getByTestId } = render(
+      <UploadProgressBar upload={instance} striped />
+    );
 
-    expect(progressBar.className).toContain('progress-bar-striped')
-    
-  })
+    const progressBar = getByTestId('progress-inner');
 
-  test('should render animated', () =>{
-    const { getByTestId } = render(<UploadProgressBar upload={instance} animated />);
-    
-    const progressBar = getByTestId('progress-inner')
+    expect(progressBar.className).toContain('progress-bar-striped');
+  });
 
-    expect(progressBar.className).toContain('progress-bar-animated')
-    
-  })
+  test('should render animated', () => {
+    const { getByTestId } = render(
+      <UploadProgressBar upload={instance} animated />
+    );
 
-  test('should render striped and animated', () =>{
-    const { getByTestId } = render(<UploadProgressBar upload={instance} striped animated />);
-    
-    const progressBar = getByTestId('progress-inner')
+    const progressBar = getByTestId('progress-inner');
 
-    expect(progressBar.className).toContain('animated')
-    expect (progressBar.className).toContain('striped')
-    
-  })
+    expect(progressBar.className).toContain('progress-bar-animated');
+  });
 
+  test('should render striped and animated', () => {
+    const { getByTestId } = render(
+      <UploadProgressBar upload={instance} striped animated />
+    );
 
+    const progressBar = getByTestId('progress-inner');
+
+    expect(progressBar.className).toContain('animated');
+    expect(progressBar.className).toContain('striped');
+  });
 });